Fixed some VCS bugs (#231)

+++ projects/vcs/pkg/git/transform.go
@@ -122,12 +122,15 @@ func (r *Repository) Import(cs *v1.ChangeSet) error {
 	 * If we just overwrite, we will not be able to represent deletions.
 	 */
 
-	if vsList, err := fileClient.VirtualServiceClient.List(defaults.GlooSystem, clients.ListOpts{}); err == nil {
-		for _, vs := range vsList {
-			fileClient.VirtualServiceClient.Delete(defaults.GlooSystem, vs.Metadata.Name, clients.DeleteOpts{IgnoreNotExist: true})
+	// We have to check if the directory exists, otherwise List will fail
+	if r.exists(constants.VirtualServiceRootDir) {
+		if vsList, err := fileClient.VirtualServiceClient.List(defaults.GlooSystem, clients.ListOpts{}); err == nil {
+			for _, vs := range vsList {
+				fileClient.VirtualServiceClient.Delete(defaults.GlooSystem, vs.Metadata.Name, clients.DeleteOpts{IgnoreNotExist: true})
+			}
+		} else {
+			return err
 		}
-	} else {
-		return err
 	}
 	for _, vs := range cs.Data.VirtualServices {
 		_, err = fileClient.VirtualServiceClient.Write(vs, clients.WriteOpts{OverwriteExisting: true})
@@ -136,88 +139,100 @@ func (r *Repository) Import(cs *v1.ChangeSet) error {
 		}
 	}
 
-	if gatewayList, err := fileClient.GatewayClient.List(defaults.GlooSystem, clients.ListOpts{}); err == nil {
-		for _, gateway := range gatewayList {
-			fileClient.GatewayClient.Delete(defaults.GlooSystem, gateway.Metadata.Name, clients.DeleteOpts{IgnoreNotExist: true})
-		}
-	} else {
-		return err
-	}
-	for _, gateway := range cs.Data.Gateways {
-		_, err = fileClient.GatewayClient.Write(gateway, clients.WriteOpts{OverwriteExisting: true})
-		if err != nil {
+	if r.exists(constants.GatewayRootDir) {
+		if gatewayList, err := fileClient.GatewayClient.List(defaults.GlooSystem, clients.ListOpts{}); err == nil {
+			for _, gateway := range gatewayList {
+				fileClient.GatewayClient.Delete(defaults.GlooSystem, gateway.Metadata.Name, clients.DeleteOpts{IgnoreNotExist: true})
+			}
+		} else {
 			return err
+		}
+		for _, gateway := range cs.Data.Gateways {
+			_, err = fileClient.GatewayClient.Write(gateway, clients.WriteOpts{OverwriteExisting: true})
+			if err != nil {
+				return err
+			}
 		}
 	}
 
-	if proxyList, err := fileClient.ProxyClient.List(defaults.GlooSystem, clients.ListOpts{}); err == nil {
-		for _, proxy := range proxyList {
-			fileClient.ProxyClient.Delete(defaults.GlooSystem, proxy.Metadata.Name, clients.DeleteOpts{IgnoreNotExist: true})
-		}
-	} else {
-		return err
-	}
-	for _, proxy := range cs.Data.Proxies {
-		_, err = fileClient.ProxyClient.Write(proxy, clients.WriteOpts{OverwriteExisting: true})
-		if err != nil {
+	if r.exists(constants.ProxyRootDir) {
+		if proxyList, err := fileClient.ProxyClient.List(defaults.GlooSystem, clients.ListOpts{}); err == nil {
+			for _, proxy := range proxyList {
+				fileClient.ProxyClient.Delete(defaults.GlooSystem, proxy.Metadata.Name, clients.DeleteOpts{IgnoreNotExist: true})
+			}
+		} else {
 			return err
+		}
+		for _, proxy := range cs.Data.Proxies {
+			_, err = fileClient.ProxyClient.Write(proxy, clients.WriteOpts{OverwriteExisting: true})
+			if err != nil {
+				return err
+			}
 		}
 	}
 
-	if resolverMapList, err := fileClient.ResolverMapClient.List(defaults.GlooSystem, clients.ListOpts{}); err == nil {
-		for _, resolverMap := range resolverMapList {
-			fileClient.ResolverMapClient.Delete(defaults.GlooSystem, resolverMap.Metadata.Name, clients.DeleteOpts{IgnoreNotExist: true})
-		}
-	} else {
-		return err
-	}
-	for _, resolverMap := range cs.Data.ResolverMaps {
-		_, err = fileClient.ResolverMapClient.Write(resolverMap, clients.WriteOpts{OverwriteExisting: true})
-		if err != nil {
+	if r.exists(constants.ResolverMapRootDir) {
+		if resolverMapList, err := fileClient.ResolverMapClient.List(defaults.GlooSystem, clients.ListOpts{}); err == nil {
+			for _, resolverMap := range resolverMapList {
+				fileClient.ResolverMapClient.Delete(defaults.GlooSystem, resolverMap.Metadata.Name, clients.DeleteOpts{IgnoreNotExist: true})
+			}
+		} else {
 			return err
+		}
+		for _, resolverMap := range cs.Data.ResolverMaps {
+			_, err = fileClient.ResolverMapClient.Write(resolverMap, clients.WriteOpts{OverwriteExisting: true})
+			if err != nil {
+				return err
+			}
 		}
 	}
 
-	if schemaList, err := fileClient.SchemaClient.List(defaults.GlooSystem, clients.ListOpts{}); err == nil {
-		for _, schema := range schemaList {
-			fileClient.SchemaClient.Delete(defaults.GlooSystem, schema.Metadata.Name, clients.DeleteOpts{IgnoreNotExist: true})
-		}
-	} else {
-		return err
-	}
-	for _, schema := range cs.Data.Schemas {
-		_, err = fileClient.SchemaClient.Write(schema, clients.WriteOpts{OverwriteExisting: true})
-		if err != nil {
+	if r.exists(constants.SchemaRootDir) {
+		if schemaList, err := fileClient.SchemaClient.List(defaults.GlooSystem, clients.ListOpts{}); err == nil {
+			for _, schema := range schemaList {
+				fileClient.SchemaClient.Delete(defaults.GlooSystem, schema.Metadata.Name, clients.DeleteOpts{IgnoreNotExist: true})
+			}
+		} else {
 			return err
+		}
+		for _, schema := range cs.Data.Schemas {
+			_, err = fileClient.SchemaClient.Write(schema, clients.WriteOpts{OverwriteExisting: true})
+			if err != nil {
+				return err
+			}
 		}
 	}
 
-	if settingsList, err := fileClient.SettingsClient.List(defaults.GlooSystem, clients.ListOpts{}); err == nil {
-		for _, setting := range settingsList {
-			fileClient.SettingsClient.Delete(defaults.GlooSystem, setting.Metadata.Name, clients.DeleteOpts{IgnoreNotExist: true})
-		}
-	} else {
-		return err
-	}
-	for _, setting := range cs.Data.Settings {
-		_, err = fileClient.SettingsClient.Write(setting, clients.WriteOpts{OverwriteExisting: true})
-		if err != nil {
+	if r.exists(constants.SettingsRootDir) {
+		if settingsList, err := fileClient.SettingsClient.List(defaults.GlooSystem, clients.ListOpts{}); err == nil {
+			for _, setting := range settingsList {
+				fileClient.SettingsClient.Delete(defaults.GlooSystem, setting.Metadata.Name, clients.DeleteOpts{IgnoreNotExist: true})
+			}
+		} else {
 			return err
+		}
+		for _, setting := range cs.Data.Settings {
+			_, err = fileClient.SettingsClient.Write(setting, clients.WriteOpts{OverwriteExisting: true})
+			if err != nil {
+				return err
+			}
 		}
 	}
 
-	if upstreamList, err := fileClient.UpstreamClient.List(defaults.GlooSystem, clients.ListOpts{}); err == nil {
-		for _, upstream := range upstreamList {
-			fileClient.UpstreamClient.Delete(defaults.GlooSystem, upstream.Metadata.Name, clients.DeleteOpts{IgnoreNotExist: true})
-		}
-	} else {
-		return err
-	}
-	// TODO: do we do upstreams?
-	for _, upstream := range cs.Data.Upstreams {
-		_, err = fileClient.UpstreamClient.Write(upstream, clients.WriteOpts{OverwriteExisting: true})
-		if err != nil {
+	if r.exists(constants.UpstreamRootDir) {
+		if upstreamList, err := fileClient.UpstreamClient.List(defaults.GlooSystem, clients.ListOpts{}); err == nil {
+			for _, upstream := range upstreamList {
+				fileClient.UpstreamClient.Delete(defaults.GlooSystem, upstream.Metadata.Name, clients.DeleteOpts{IgnoreNotExist: true})
+			}
+		} else {
 			return err
+		}
+		// TODO: do we do upstreams?
+		for _, upstream := range cs.Data.Upstreams {
+			_, err = fileClient.UpstreamClient.Write(upstream, clients.WriteOpts{OverwriteExisting: true})
+			if err != nil {
+				return err
+			}
 		}
 	}
